requirements.txt
ossapi==5.2.1
dotenv==0.9.9
discord.py==2.5.2
Flask==3.1.1
Flask-Cors==4.0.0
python-dateutil==2.9.0.post0
numpy==2.2.6
osu-db-tools==0.1.0
apscheduler=3.11.0
redis=6.4.0
